Fix initialization under reduced privileges
Also:

* fixed timed log rotation
* --log-level option added

+++ snmpsim/daemon.py
@@ -7,6 +7,7 @@
 import sys
 from snmpsim import error
 
+
 if sys.platform[:3] == 'win':
     def daemonize(pidfile):
         raise error.SnmpsimError('Windows is not inhabited with daemons!')
@@ -14,8 +15,14 @@ if sys.platform[:3] == 'win':
 
     def dropPrivileges(uname, gname):
         return
+
 else:
-    import os, pwd, grp, atexit, signal, tempfile
+    import os
+    import pwd
+    import grp
+    import atexit
+    import signal
+    import tempfile
 
 
     def daemonize(pidfile):
@@ -53,19 +60,22 @@ else:
         # write pidfile
         def atexit_cb():
             try:
-                os.remove(pidfile)
+                if pidfile:
+                    os.remove(pidfile)
             except OSError:
                 pass
 
         atexit.register(atexit_cb)
 
         try:
-            fd, nm = tempfile.mkstemp(dir=os.path.dirname(pidfile))
-            os.write(fd, ('%d\n' % os.getpid()).encode('utf-8'))
-            os.close(fd)
-            os.rename(nm, pidfile)
-        except:
-            raise error.SnmpsimError('Failed to create PID file %s: %s' % (pidfile, sys.exc_info()[1]))
+            if pidfile:
+                fd, nm = tempfile.mkstemp(dir=os.path.dirname(pidfile))
+                os.write(fd, ('%d\n' % os.getpid()).encode('utf-8'))
+                os.close(fd)
+                os.rename(nm, pidfile)
+        except Exception:
+            raise error.SnmpsimError(
+                'Failed to create PID file %s: %s' % (pidfile, sys.exc_info()[1]))
 
         # redirect standard file descriptors
         sys.stdout.flush()
@@ -79,32 +89,69 @@ else:
         os.dup2(se.fileno(), sys.stderr.fileno())
 
 
-    def dropPrivileges(uname, gname):
-        if os.getuid() != 0:
-            if uname and uname != pwd.getpwnam(uname).pw_name or \
-                            gname and gname != grp.getgrnam(gname).gr_name:
-                raise error.SnmpsimError('Process is running under different UID/GID')
+    class PrivilegesOf(object):
+
+        def __init__(self, uname, gname, final=False):
+            self._uname = uname
+            self._gname = gname
+            self._final = final
+            self._olduid = self._oldgid = None
+
+        def __enter__(self):
+            if os.getuid() != 0:
+                if (self._uname and self._uname != pwd.getpwnam(self._uname).pw_name or
+                        self._gname and self._gname != grp.getgrnam(self._gname).gr_name):
+                    raise error.SnmpsimError('Process is running under different UID/GID')
+                else:
+                    return
             else:
+                if not self._uname or not self._gname:
+                    raise error.SnmpsimError('Must drop privileges to a non-privileged user&group')
+
+            try:
+                runningUid = pwd.getpwnam(self._uname).pw_uid
+                runningGid = grp.getgrnam(self._gname).gr_gid
+
+            except Exception:
+                raise error.SnmpsimError(
+                    'getpwnam()/getgrnam() failed for %s/%s: %s' % (
+                        self._uname, self._gname, sys.exc_info()[1]))
+
+            try:
+                os.setgroups([])
+
+            except Exception:
+                raise error.SnmpsimError('setgroups() failed: %s' % sys.exc_info()[1])
+
+            try:
+                if self._final:
+                    os.setgid(runningGid)
+                    os.setuid(runningUid)
+
+                else:
+                    self._olduid = os.getuid()
+                    self._oldgid = os.getgid()
+
+                    os.setegid(runningGid)
+                    os.seteuid(runningUid)
+
+            except Exception:
+                raise error.SnmpsimError(
+                    '%s failed for %s/%s: %s' % (
+                        self._final and 'setgid()/setuid()' or 'setegid()/seteuid()',
+                        runningGid, runningUid, sys.exc_info()[1]))
+
+            os.umask(63)  # 0077
+
+        def __exit__(self, *args):
+            if self._olduid is None or self._oldgid is None:
                 return
-        else:
-            if not uname or not gname:
-                raise error.SnmpsimError('Must drop priveleges to a non-priveleged user&group')
 
-        try:
-            runningUid = pwd.getpwnam(uname).pw_uid
-            runningGid = grp.getgrnam(gname).gr_gid
-        except Exception:
-            raise error.SnmpsimError('getpwnam()/getgrnam() failed for %s/%s: %s' % (uname, gname, sys.exc_info()[1]))
+            try:
+                os.setegid(self._oldgid)
+                os.seteuid(self._olduid)
 
-        try:
-            os.setgroups([])
-        except Exception:
-            raise error.SnmpsimError('setgroups() failed: %s' % sys.exc_info()[1])
-
-        try:
-            os.setgid(runningGid)
-            os.setuid(runningUid)
-        except Exception:
-            raise error.SnmpsimError('setgid()/setuid() failed for %s/%s: %s' % (runningGid, runningUid, sys.exc_info()[1]))
-
-        os.umask(63)  # 0077
+            except Exception:
+                raise error.SnmpsimError(
+                    'setegid()/seteuid() failed for %s/%s: %s' % (
+                        self._oldgid, self._olduid, sys.exc_info()[1]))
